refactor(nms): Move babel-register out of fbc_js_core
