+++ nms/app/common/polyfill.js
