+++ nms/babelRegister.js
