+++ nms/nodemon.json
@@ -7,5 +7,5 @@
     "grafana/",
     "alerts/"
   ],
-  "exec": "node -r './fbc_js_core/babel_register/index.js'"
+  "exec": "node -r './babelRegister.js'"
 }
